+++ src/ActionGameCore/Assets/CharacterControls/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [Development]
+
+### Changed
+
+- Rename CharacterMoveInput to CharacterMoveInputRelay.
+
+### Added
+
+- Added bot input samples.
+
+### Modified
+
+- Merge CharacterControls assembly definitions.
+
 ## [0.12.0]
 
 ### Changed